move diffsl code to new file
